api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Ahmad <ahmad.hussain.307@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
celery = {extras = ["redis"], version = "^5.3.6"}
uuid = "^1.30"
pytest = "^8.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
